+++ internal/diags/problems.go
@@ -6,6 +6,8 @@ import (
 	"slices"
 	"strings"
 
+	promParser "github.com/prometheus/prometheus/promql/parser"
+
 	"github.com/cloudflare/pint/internal/output"
 )
 
@@ -81,6 +83,7 @@ const (
 )
 
 type Diagnostic struct {
+	Expr        promParser.Node `yaml:"-"`
 	Message     string
 	Pos         PositionRanges `yaml:"-"`
 	FirstColumn int            // 1-indexed